requirements.txt
tqdm==4.67.1
colorama==0.4.6
rich==13.9.4
mysql-connector-python==9.1.0
pyfiglet==1.0.2
inquirer==3.4.0
